Added Files
Encode an English text with the given offset.
public static String encode(int offset, String original) {
...
}
If the offset is 1 then all 'a' letters are replaced with 'b', 'b' with 'c', etc.
If the offset is -1 then all 'a' letters are replaced with 'z', 'b' with 'a', etc.
The task must be implemented in any version of Java. All non-character symbols are not modified.